This commit was generated by cvs2svn to compensate for changes in r15369, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@15370 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/drive/drivecpu.c
@@ -366,7 +366,8 @@ inline static DWORD drive_trap_handler(drive_context_t *drv)
         MOS6510_REGS_SET_PC(&(drv->cpu->cpu_regs), 0xdaf6);
         return 0;
     }
-    return -1;
+
+    return (DWORD)-1;
 }
 
 static void drive_generic_dma(void)
